requirements.txt
pymongo[srv]>=4.6,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
xmltodict>=0.13,<1.0
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
tenacity>=8.2,<10.0
structlog>=24.1,<26.0
pydantic>=2.0,<3.0
redis>=5.0,<6.0
celery>=5.4,<6.0
polars>=1.0,<2.0
sse-starlette>=2.0,<3.0
